Fixes custom.css caching problem
I had tried a hack where all files in assets/ were marked as static/ to force them to get copied to the output. Unfortunately that meant that custom.css was being cached and never updated on livereload which made it useless to do work with.

I've now changed asset managment strategy to split things up so that things that must be published (without hugo knowing about it) are manully copied into static/, This is a sub-optimal setup but until Hugo supports wildcard matching and copying in it's assets (pipes) it'll have to do.