Fixes annotorious canvas resize bug
Annotorious is a right pain. When the browser is resized after anno has initiated, annortorious' canvas will not resize and will overflow.

This hack listens for browser resize events and changes the zoom of the annotorious container to be relative to the desired size. Changing the zoom affects all components in side and their coordinate systems, which is why it works - as opposed to any other solution, e.g. setting width/height, or transform scales etc...