Massive overhaul of annotate audio image.
Apparently the CSS zoom property is an IE-ism which Chrome supports but FF does not :angry:

I've rebuilt the drawing surface so it is scalable with Css transforms. It now draws boxes well, without position bugs due to the zoom, in both Chrome and Firefox. Warning though: this WILL NOT WORK on touch devices because annotorious is outdated.

Also fixed timecode inconsistent formatting that caused the padded margins to change.

Also removed RHS duration time code so that the image has more width available.

Also let's laptops with even smaller screens in because we have better support for scaling.