Enhancements to likert width
Builds off existing change to limit likert width. Now ensures width is the same for both jspsych and custom likert scales, and will reach a maximum width for wider screens.

Also scaled radio inputs to they are easier to hit.

The new changes are also moved to a better spot; our stylesheet is a better place to make changes. Making changes to the jspsych content files could result in those changes being overwritten if we update the source files.