Fixed playback cursor not moving on Edge
It turns out, the "progress" event is not the correct event to bind to for media position changes. Progress events are meant to be raised when resources are loaded (which has different meanings in different browsers). In Chrome and FF "progress" is fired regularly, in Edge it is never fired.

The "timeupdate" event is correct event to bind to. Tested on FF, Chrome, and Edge.